+++ ArduinoJson.h
@@ -1,0 +1,14 @@
+// Copyright Benoit Blanchon 2014-2015
+// MIT License
+//
+// Arduino JSON library
+// https://github.com/bblanchon/ArduinoJson
+
+// About this file
+// ---------------
+// This file is here to please the platform.io compile. It must be present in
+// the root for the IDE to find it. Feel free to ignore this file if your
+// working in another environment.
+// ...and yes this is driving me crazy!
+
+#include "include/ArduinoJson.h"